public/reference-impl/requirements.txt
# AAP Reference Implementation Requirements

# JWT handling
PyJWT==2.8.0
cryptography==41.0.7

# JSON Schema validation
jsonschema==4.20.0

# HTTP server (for AS endpoints)
Flask==3.0.0
Werkzeug==3.0.1

# HTTP client (for testing RS)
requests==2.31.0

# Date/time handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
